+++ changelog.md
@@ -3,6 +3,10 @@
 * Updated .toc for patch 8.1
 * Added patch 8.1 as an option for Loot History mass deletion.
 
+* **Auto Trade**
+* Added a new option that when enabled bypasses the trade popup and automatically adds items that should be traded.
+* This option is disabled by default.
+
 ###### Bugfixes
 * *Fixed issue with fewer than default buttons not working properly (Git#172).*
 * *Items would sometimes be marked as non-tradeable if the player already had a duplicate of the item (#379, #375).*